Minor stuff

+++ src/render/RenderEntity.cpp
@@ -4,11 +4,11 @@
 
 #include "RenderEntity.h"
 
-RenderEntity::RenderEntity(Rectangle bounds, double angle) : bounds(bounds), angle(angle) {}
+RenderEntity::RenderEntity(Rectangle bounds, float angle) : bounds(bounds), angle(angle) {}
 
 RenderEntity::RenderEntity(Rectangle bounds) : RenderEntity(bounds, 0) {}
 
-RenderEntity::RenderEntity(double x, double y, double width, double height, double angle)
+RenderEntity::RenderEntity(double x, double y, double width, double height, float angle)
         : RenderEntity({x, y, width, height}, angle) {}
 
 const Rectangle &RenderEntity::getBounds() const {
@@ -19,10 +19,10 @@ void RenderEntity::setBounds(const Rectangle &bounds) {
     RenderEntity::bounds = bounds;
 }
 
-double RenderEntity::getAngle() const {
+float RenderEntity::getAngle() const {
     return angle;
 }
 
-void RenderEntity::setAngle(double angle) {
+void RenderEntity::setAngle(float angle) {
     RenderEntity::angle = angle;
 }
